requirements.txt
filetype==1.2.0
fastapi==0.73.0
fastapi-utils==0.2.1
python-multipart==0.0.6
uvicorn[standard]==0.17.6
PyJWT==2.3.0
urllib3==1.26.9
passlib==1.7.4
pymongo~=4.0.1
jdatetime
email-validator
starlette~=0.17.1
pydantic~=1.9.1
minio~=7.1.17
python-dotenv~=0.19.2
setuptools~=60.2.0